requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==1.10.19
pyjwt==2.10.0
passlib[bcrypt]==1.7.4
